[WebAssembly][CodeGen] Allow for externref/funcref local variables
Add logic to instruction selection to lower static alloca of pointers in
certain well-known non-default address spaces to WebAssembly local
variables.

The motivation is that many language front-ends will want to represent
local variables as alloca's, relying on SROA / mem2reg to lift those
allocations to SSA variables as appropriate.  If optimizations are off,
this won't happen, and those alloca's will reach instruction selection.
But externref and funcref values can't be written to linear memory, so
there we have a problem.

The solution we are going to work towards is to annotate the initial
alloca's of reference-typed values with a special address space.
Pointers in this address space denote automatic-storage-duration
locations, i.e., function-local variables.  The instruction selector
sees loads and stores from these locations and turns the allocations
into WebAssembly locals, and turns the loads and stores into LOCAL_GET /
LOCAL_SET.