[IR][Verifier] Allow alloca outside alloca address space
The WebAssembly target has a class of values that can't be written to
the stack, but would like to allow function-local variables to be
modelled using the usual alloca approach.  These allocations will be
lowered to off-stack automatic-duration storage locations during
instruction selection.  For this strategy to work, we need to relax the
restriction on alloca IR instructions.