[WebAssembly] Renumber reference-types address spaces
For externref, we currently use two address spaces: one for externref
values, and one for externref locations (global variables).  We use two
additional address spaces for funcrefs.  In the future we'll need one
more address space each, for access via local variables.  But then it
gets hard to remember which is which.

So, instead of treating address spaces as a scarce commodity, this patch
changes to have the different variants grouped contiguously, with some
slace space if we decide more address spaces are needed.