[WebAssembly] Use address space for WebAssembly globals
Firstly, this patch makes TargetLowering::getPointerMemTy virtual, as
getPointerTy is.  Otherwise, LLVM will conclude that the type of an
externref in memory is void*, and the ISD::LOAD node for externref will
have the wrong type (i32 instead of externref).

Before this change, the only way to know whether a load was for an
externref vs a funcref was to look at the address space of the base
pointer, so globals of different types needed different address spaces.
Now the loaded type is correctly computed, so we can instead match on
the type of the ISD::LOAD result, and just use one address space for
globals of any type.