[WebAssembly] Add HeapType, RefType definitions; use for tables etc
The misleadingly-named [typed function
references](https://github.com/WebAssembly/function-references) proposal
is a precursor to the garbage collection proposal, as well as
stringrefs.

This patches updates LLVM and LLD to have HeapType and RefType type
definitions, corresponding to the `heaptype` and `reftype` definitions
introduced in the specification.  The emitted code should be exactly the
same, except that the YAML interface changes to already have the more
expressive capability to refer to types by index, to declare
non-nullable types, and so on.