[WebAssembly] Add stringref data types
The [stringref](https://github.com/WebAssembly/stringref) proposal is
based on typed function references and provides reference-typed,
host-managed strings.  Along with a number of generic string operations,
one can additionally request a view on the string in terms of WTF-8 (a
UTF-8 variant), WTF-16 (same as in Java or JavaScript), or a codepoint
iteration view.

This patch adds the basics of support for stringrefs.  No tests yet
though!